fix: 自己紹介取得はgetProfile内であるため、フックの名前をuseGetProfileに修正
